pyproject.toml
[project]
name = "todotoday"
version = "0.1.0"
description = "Daily briefing agent using Claude and Apple ecosystem tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk",
    "fastmcp>=3.0",
    "python-dotenv",
    "pyobjc-framework-EventKit",
]

[project.scripts]
todotoday = "todotoday.run_briefing:main"

[tool.uv]
package = true

[tool.uv.sources]
claude-agent-sdk = { git = "https://github.com/anthropics/claude-agent-sdk-python" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/todotoday"]
